+++ tests/Commands/__snapshots__/JobMakeCommandTest__it_generated_correct_sync_job_file_with_content__1.php
@@ -1,16 +1,13 @@
 <?php return '<?php
 
-namespace DummyNamespace;
+namespace Modules\\Blog\\Jobs;
 
 use Illuminate\\Bus\\Queueable;
-use Illuminate\\Queue\\SerializesModels;
-use Illuminate\\Queue\\InteractsWithQueue;
-use Illuminate\\Contracts\\Queue\\ShouldQueue;
 use Illuminate\\Foundation\\Bus\\Dispatchable;
 
-class DummyClass implements ShouldQueue
+class SomeJob implements ShouldQueue
 {
-    use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
+    use Dispatchable, Queueable;
 
     /**
      * Create a new job instance.